made scatter enemy that inherits from superclass
